op3: livedisplay: Use modules from common implementation
