--- lineagehw/livedisplay/SunlightEnhancement.cpp
@@ -1,67 +1,0 @@
-/*
- * Copyright (C) 2019-2020 The LineageOS Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#define LOG_TAG "vendor.lineage.livedisplay@2.0-impl.oneplus3"
-
-#include "SunlightEnhancement.h"
-
-#include <android-base/file.h>
-#include <android-base/logging.h>
-#include <android-base/strings.h>
-
-using ::android::base::ReadFileToString;
-using ::android::base::Trim;
-using ::android::base::WriteStringToFile;
-
-namespace {
-constexpr char kFileHBM[] = "/sys/class/graphics/fb0/hbm";
-}  // anonymous namespace
-
-namespace vendor {
-namespace lineage {
-namespace livedisplay {
-namespace V2_0 {
-namespace sysfs {
-
-bool SunlightEnhancement::isSupported() {
-    bool supported = !access(kFileHBM, R_OK | W_OK);
-    if (!supported) {
-        LOG(ERROR) << "AdaptiveBacklight not supported!";
-    }
-    return supported;
-}
-
-// Methods from ::vendor::lineage::livedisplay::V2_0::ISunlightEnhancement follow.
-Return<bool> SunlightEnhancement::isEnabled() {
-    std::string tmp;
-    int32_t contents = 0;
-
-    if (ReadFileToString(kFileHBM, &tmp)) {
-        contents = std::stoi(Trim(tmp));
-    }
-
-    return contents > 0;
-}
-
-Return<bool> SunlightEnhancement::setEnabled(bool enabled) {
-    return WriteStringToFile(enabled ? "1" : "0", kFileHBM, true);
-}
-
-}  // namespace sysfs
-}  // namespace V2_0
-}  // namespace livedisplay
-}  // namespace lineage
-}  // namespace vendor
