+++ lineagehw/livedisplay/service.cpp
@@ -14,20 +14,18 @@
  * limitations under the License.
  */
 
-
 #define LOG_TAG "vendor.lineage.livedisplay@2.0-service.oneplus3"
 
 #include <android-base/logging.h>
 #include <binder/ProcessState.h>
 #include <hidl/HidlTransportSupport.h>
+#include <livedisplay/sdm/AdaptiveBacklight.h>
+#include <livedisplay/sdm/PictureAdjustment.h>
+#include <livedisplay/sysfs/SunlightEnhancement.h>
 
 #include <functional>
 
-#include "AdaptiveBacklight.h"
-#include "DisplayModes.h"
-#include "PictureAdjustment.h"
-#include "SDMController.h"
-#include "SunlightEnhancement.h"
+#include "MixedDisplayModes.h"
 
 using ::android::OK;
 using ::android::sp;
@@ -35,32 +33,33 @@ using ::android::hardware::configureRpcThreadpool;
 using ::android::hardware::joinRpcThreadpool;
 
 using ::vendor::lineage::livedisplay::V2_0::sdm::AdaptiveBacklight;
-using ::vendor::lineage::livedisplay::V2_0::sdm::DisplayModes;
+using ::vendor::lineage::livedisplay::V2_0::sdm::MixedDisplayModes;
 using ::vendor::lineage::livedisplay::V2_0::sdm::PictureAdjustment;
 using ::vendor::lineage::livedisplay::V2_0::sdm::SDMController;
 using ::vendor::lineage::livedisplay::V2_0::sysfs::SunlightEnhancement;
 
 bool RegisterSdmServices() {
     std::shared_ptr<SDMController> controller = std::make_shared<SDMController>();
-    sp<DisplayModes> dm = new DisplayModes(controller);
+    sp<MixedDisplayModes> dm = new MixedDisplayModes(controller);
     sp<PictureAdjustment> pa = new PictureAdjustment(controller);
 
     // Update default PA on setDisplayMode
-    dm->registerCb(std::bind(&PictureAdjustment::updateDefaultPictureAdjustment, pa));
+    dm->registerDisplayModeSetCallback(
+            std::bind(&PictureAdjustment::updateDefaultPictureAdjustment, pa));
 
     return dm->registerAsService() == OK && pa->registerAsService() == OK;
 }
 
 bool RegisterSysfsServices() {
-    if (AdaptiveBacklight::isSupported()) {
-        sp<AdaptiveBacklight> ab = new AdaptiveBacklight();
+    sp<AdaptiveBacklight> ab = new AdaptiveBacklight();
+    if (ab->isSupported()) {
         if (ab->registerAsService() != OK) {
             return false;
         }
     }
 
-    if (SunlightEnhancement::isSupported()) {
-        sp<SunlightEnhancement> se = new SunlightEnhancement();
+    sp<SunlightEnhancement> se = new SunlightEnhancement();
+    if (se->isSupported()) {
         if (se->registerAsService() != OK) {
             return false;
         }
